Crate ic_registry_routing_table

Structs

CanisterIdRange

CanisterIdRanges
A list of closed CanisterId ranges that are present in the RoutingTable

RoutingTable
Stores an ordered map mapping CanisterId ranges to SubnetIds. The ranges tracked are inclusive of start and end i.e. can be denoted as [a, b].

Enums

ResolveDestinationError

WellFormedError

Functions

resolve_destination
Inspect the method name and payload of a request to ic:00 to figure out to which subnet it should be sent to.

routing_table_insert_subnet
A helper function to help insert a new subnet to the routing table