Struct ic_registry_routing_table::RoutingTable

pub struct RoutingTable(pub BTreeMap<CanisterIdRange, SubnetId>);

Stores an ordered map mapping CanisterId ranges to SubnetIds. The ranges tracked are inclusive of start and end i.e. can be denoted as [a, b].

Implementations

impl RoutingTable

pub fn new(map: BTreeMap<CanisterIdRange, SubnetId>) -> Self

pub fn insert(
&mut self,
canister_id_range: CanisterIdRange,
subnet_id: SubnetId
) -> Result<(), WellFormedError>

Insert a new subnet with a corresponding canister_id_range into the routing table. You may want to consider using routing_table_insert_subnet instead. It may meet your needs.

pub fn iter(&self) -> impl Iterator<Item = (&CanisterIdRange, &SubnetId)>

pub fn is_empty(&self) -> bool

pub fn well_formed(&self) -> Result<(), WellFormedError>

Returns true if the routing table is well-formed.

pub fn route(&self, principal_id: PrincipalId) -> Option<SubnetId>

Returns the SubnetId that the given principal_id is assigned to or None if an assignment cannot be found.

pub fn ranges(&self, subnet_id: SubnetId) -> CanisterIdRanges

Find all canister ranges that are assigned to subnet_id.